requirements.txt
flask
openai
twilio
python-dotenv
pandas
sentence-transformers
scikit-learn 